refactor(app): 我中增加关于。把当前版本收到关于中
