+++ src/app/src/screens/about/VersionScreen.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Text, Linking, Image } from 'react-native';
+import { Text, Linking } from 'react-native';
 import config from '@src/shared/project.config';
 import { WingBlank } from '@ant-design/react-native';
-import { TButton } from '../components/TComponent';
+import { TButton } from '../../components/TComponent';
 import styled from 'styled-components/native';
 import semver from 'semver';
 import _isNil from 'lodash/isNil';
